refactor(langchain): move `_declare_elicitation_capability` to `elicitation`
The sentinel handler exists purely to make the client advertise the
elicitation capability, and every line of its rationale is about elicitation —
it sat in `adapter` only because that is where the client is constructed.

`elicitation` is the package's leaf module, so importing it from `adapter`
introduces no cycle.
